requirements.txt
langchain-openai
langchain
langchain_community
python-dotenv
streamlit
ipykernel
pypdf
beautifulsoup4
pymupdf
wikipedia
langchain-text-splitters
sentence_transformers
langchain_huggingface
faiss-cpu
duckdb
pandas
openai
langchain_groq
duckduckgo-search
mysql-connector-python
SQLAlchemy
flask-cors

